Minor clean-ups, couple of small liquidator features (#58)
* Removed redundant function filler.sleep, using utils.sleepMs

Minor clean-up. Not sure why the filler bot added its own sleep
function when there was one available.

* Fixed minor comment typo on filler bot

* Using sleepMs on fillerLite

It was using the inherited `sleep` from Filler, but I'd removed it on
6ffbc83.

* Added comment on example config about market definitions

New users wouldn't necessarily know where to look otherwise.

* Feature: Added minimum liquidation amount option

This value allows you to skip accounts which have too small amount to
be worth attempting to liquidate. I've found that liquidation often
fails with accounts that have a couple of lamports left, no positions,
and an unclaimed negative PnL.

* Updating SDK to the latest

* Feature: liquidator can now ignore some accounts

I was getting persistent logs when skipping or attempting to liquidate
what seems to be un-liquidatable accounts. This feature allows the
liquidator to skip them altogether to avoid noise.

* Fixed sample config name on README.

* Added config example for excluded liquidator accounts

The main change is on commit e8ad5e6, this acts as an example for how
to list them.

* BUGFIX: Liquidator spotMarketIndicies were empty on subaccount config

Liquidator was initializing spotMarketIndicies only if there was no
subaccount configuration - otherwise it did not get assigned.

* Enhancement: not acquiring mutex for liquidator when single account

tryLiquidateStart acquires the mutex since both that function and the
derisking process may switch subaccounts. This is not necessary on a
single account, so no need to block while derisking is taking place.

* Enhancement: ignoring acccounts while getting usersCanBeLiquidated

Moving filtering up, and extracting userKey as we go through the
accounts that can be liquidated.

Follows up commit 1be4cb9.

See:

https://github.com/drift-labs/keeper-bots-v2/pull/58#pullrequestreview-1590330502

* Using BN number/string conversion

I was comparing them directly to a number, but ltn seems to be the right
way to go about it.

* Updated SDK to 2.38.0 to avoid collateral calculation issues

This also required updating jit-proxy. We cannot go with the beta
version from commit 3733b01, or wilt get a version conflict.

See https://github.com/orgs/drift-labs/discussions/594

* Filtering spot amounts below config threshold on findBestSpotPosition

This helps us get right of positions we do not want to clear out
earlier than before.

Improved comment on example to show this value is specified on
lamports.

See https://github.com/drift-labs/keeper-bots-v2/pull/58#discussion_r1309213465

* Clean-up: building perpMarketIndexes like in ee823fd

See https://github.com/drift-labs/keeper-bots-v2/pull/58/files#r1303833783

* Minor comments, expanded README

This commit addresses the message discussed on issue #62.

* Addressing merge conflict on package.json

* Addressing liquidator comments from PR #58

- scaledBalance can be negative in case of borrows
- Should use ZERO with minDepositToLiq, since it's a BN