floatingMaker handling a "closing" state
Decided to handle what happens if we get too lopsided, and we need to
start closing one side. We identify that case by the proportion being
such  that one of the sides would be smaller than the minimum order
size for the market. In that case, we *will* allow a single order on
the other side.